menambahkan fitur ubah data

+++ index.php
@@ -23,10 +23,10 @@ $warkop = query("SELECT * FROM menu");
     <tr>
       <th>#</th>
       <th>Nama</th>
+      <th>Gambar</th>
       <th>Alamat</th>
       <th>Makanan</th>
       <th>Minuman</th>
-      <th>Gambar</th>
       <th>Aksi</th>
     </tr>
 
@@ -35,13 +35,13 @@ $warkop = query("SELECT * FROM menu");
       <tr>
         <td><?= $i++; ?></td>
         <td><?= $w['tempat']; ?></td>
+        <td><img width="100" src="img/<?= $w['gambar']; ?>" alt=""></td>
         <td><?= $w['alamat']; ?></td>
         <td><?= $w['makanan']; ?></td>
         <td><?= $w['minuman']; ?></td>
-        <td><img width="100" src="img/<?= $w['gambar']; ?>" alt=""></td>
         <td>
-          <a href="">Ubah</a> | 
-          <a href="hapus.php?id=<?= $w['id']; ?>">Delete</a>
+          <a href="ubah.php?id=<?= $w['id']; ?>">Ubah</a> | 
+          <a href="hapus.php?id=<?= $w['id']; ?>" onclick="return confirm('Apakah Anda yakin?')">Delete</a>
         </td>
       </tr>
     <?php endforeach; ?>
